⚡️ support pattern result matching by walking analyses outdirs (#71)
* ⚡️ support pattern result matching by walk outdirs
* ✅ remove edge cases from cov test
* 🤖 fix typo in docstring

+++ tests/test-with-compose.sh
@@ -1,3 +1,14 @@
+# parse arguments
+SKIP_BUILD=false
+for arg in "$@"; do
+  case $arg in
+    --skip-build)
+    SKIP_BUILD=true
+    shift
+    ;;
+  esac
+done
+
 # get path to demo directory
 if [ "$SHELL" = "zsh" ]; then
    TEST_DIR="$( cd "$(dirname ${(%):-%N})" ; pwd -P )"
@@ -15,14 +26,22 @@ echo "CLI directory set to: $CLI_DIR"
 ISABL_BRANCH=${TRAVIS_PULL_REQUEST_BRANCH:-${TRAVIS_BRANCH:-master}}
 echo "ISABL_BRANCH set to $ISABL_BRANCH given travis branch: $TRAVIS_BRANCH $TRAVIS_PULL_REQUEST_BRANCH"
 
-# clone api from github
-rm -rf $API_DIR && git clone https://github.com/papaemmelab/isabl_api.git $API_DIR
-cd $API_DIR && (git checkout $ISABL_BRANCH || true) && docker-compose build && docker-compose up -d
-
-# give some time to API to start and test Isabl CLI
-echo "Giving 30 seconds to API to get started..."
+# reset environment variables
+unset ISABL_CLIENT_ID
+unset ISABL_API_URL
 export ISABL_CLIENT_ID=test-cli-client
-sleep 30 && cd $CLI_DIR && pytest -vs tests/ --cov=isabl_cli || (
+
+if [ "$SKIP_BUILD" = false ]; then
+   # clone api from github
+   rm -rf $API_DIR && git clone https://github.com/papaemmelab/isabl_api.git $API_DIR
+   cd $API_DIR && (git checkout $ISABL_BRANCH || true) && docker-compose build && docker-compose up -d
+
+   # give some time to API to start and test Isabl CLI
+   echo "Giving 30 seconds to API to get started..."
+   sleep 30
+fi
+
+cd $CLI_DIR && pytest -vs tests/ --cov=isabl_cli || (
    # print django logs if tests fail
    cd $API_DIR && docker-compose logs django && exit 1
 )
